// src/theme/typography/types.ts
import { TypographyOptions } from "@mui/material/styles/createTypography";

export interface TypographyVariantsOptions extends TypographyOptions {
  caps26: React.CSSProperties;
  caps24: React.CSSProperties;
  caps16: React.CSSProperties;
  caps14: React.CSSProperties;
  caps12: React.CSSProperties;
  text24: React.CSSProperties;
  text22: React.CSSProperties;
  text20: React.CSSProperties;
  text18: React.CSSProperties;
  text16: React.CSSProperties;
  text14: React.CSSProperties;
  text12: React.CSSProperties;
  text10: React.CSSProperties;
  text14Medium: React.CSSProperties;
  text16Medium: React.CSSProperties;
  text18Medium: React.CSSProperties;
  text20Medium: React.CSSProperties;
  text24Medium: React.CSSProperties;
  text24Book: React.CSSProperties;
  text22Book: React.CSSProperties;
  text20Book: React.CSSProperties;
  text18Book: React.CSSProperties;
  text16Book: React.CSSProperties;
  text14Book: React.CSSProperties;
  text12Book: React.CSSProperties;
  text10Book: React.CSSProperties;
}

declare module "@mui/material/Typography" {
  interface TypographyPropsVariantOverrides {
    caps26: true;
    caps24: true;
    caps16: true;
    caps14: true;
    caps12: true;
    text24: true;
    text22: true;
    text20: true;
    text18: true;
    text16: true;
    text14: true;
    text12: true;
    text10: true;
    text12Medium: true;
    text14Medium: true;
    text16Medium: true;
    text18Medium: true;
    text20Medium: true;
    text24Medium: true;
    text24Book: true;
    text22Book: true;
    text20Book: true;
    text18Book: true;
    text16Book: true;
    text14Book: true;
    text12Book: true;
    text10Book: true;
  }
}

declare module "@mui/material/styles/createTypography" {
  interface Typography {
    caps26: React.CSSProperties;
    caps24: React.CSSProperties;
    caps16: React.CSSProperties;
    caps14: React.CSSProperties;
    caps12: React.CSSProperties;
    text24: React.CSSProperties;
    text22: React.CSSProperties;
    text20: React.CSSProperties;
    text18: React.CSSProperties;
    text16: React.CSSProperties;
    text14: React.CSSProperties;
    text12: React.CSSProperties;
    text10: React.CSSProperties;
    text14Medium: React.CSSProperties;
    text16Medium: React.CSSProperties;
    text18Medium: React.CSSProperties;
    text20Medium: React.CSSProperties;
    text24Medium: React.CSSProperties;
    text24Book: React.CSSProperties;
    text22Book: React.CSSProperties;
    text20Book: React.CSSProperties;
    text18Book: React.CSSProperties;
    text16Book: React.CSSProperties;
    text14Book: React.CSSProperties;
    text12Book: React.CSSProperties;
    text10Book: React.CSSProperties;
  }

  interface TypographyOptions {
    caps26: React.CSSProperties;
    caps24: React.CSSProperties;
    caps16: React.CSSProperties;
    caps14: React.CSSProperties;
    caps12: React.CSSProperties;
    text24: React.CSSProperties;
    text22: React.CSSProperties;
    text20: React.CSSProperties;
    text18: React.CSSProperties;
    text16: React.CSSProperties;
    text14: React.CSSProperties;
    text12: React.CSSProperties;
    text10: React.CSSProperties;
    text14Medium: React.CSSProperties;
    text16Medium: React.CSSProperties;
    text18Medium: React.CSSProperties;
    text20Medium: React.CSSProperties;
    text24Medium: React.CSSProperties;
    text24Book: React.CSSProperties;
    text22Book: React.CSSProperties;
    text20Book: React.CSSProperties;
    text18Book: React.CSSProperties;
    text16Book: React.CSSProperties;
    text14Book: React.CSSProperties;
    text12Book: React.CSSProperties;
    text10Book: React.CSSProperties;
  }
}